removed non-flat APIs
