--- algorithm/flat/internal/line_intersector/nonrobust_line_intersector_test.go
@@ -1,14 +1,0 @@
-package line_intersector_test
-
-import (
-	"github.com/twpayne/go-geom/algorithm/flat/internal/line_intersector"
-	"testing"
-)
-
-func TestNonRobustLineIntersectionPointOnLine(t *testing.T) {
-	exectuteLineIntersectionPointOnLineTest(t, line_intersector.NonRobustLineIntersector{})
-}
-
-func TestNonRobustLineIntersectionLines(t *testing.T) {
-	executeLineIntersectionLinesTest(t, line_intersector.NonRobustLineIntersector{})
-}
